+++ Assets/Scripts/Models/API/Websocket/General/Responses/WebsocketGetFeeResponse.cs
@@ -6,6 +6,6 @@ namespace General
     public class WebsocketGetFeeResponse : WebsocketGeneralResponse
     {
         public string symbol;
-        public double fee;
+        public float fee;
     }
 }
