+++ Assets/Scripts/Models/API/Websocket/General/Responses/WebsocketGetRuntimeDataResponse.cs
@@ -20,10 +20,10 @@ namespace General
         public string symbol;
         public CalculateMargin marginCalculator;
         public OrderTypeEnum orderType;
-        public double quantityFilled;
-        public double averagePriceFilled;
-        public double actualTakeProfitPrice;
-        public double paidFundingAmount;
+        public float quantityFilled;
+        public float averagePriceFilled;
+        public float actualTakeProfitPrice;
+        public float paidFundingAmount;
         public List<WebsocketGetThrottleOrderResponse> throttleOrders;
     }
     [Serializable]
@@ -39,7 +39,7 @@ namespace General
     public class WebsocketGetQuickOrderResponse
     {
         public string id;
-        public double entryPrice;
+        public float entryPrice;
         public bool isLong;
         public Preference setting;
     }
